QI.9v5kg2b: More refactoring
Pulling out that actual implementation of the HistoryFolding.
